+++ lib/can_be/processor/instance.rb
@@ -5,6 +5,9 @@ module CanBe
         @model = model
         @config = model.class.can_be_config
         @field_name = @config.field_name
+        @details_name = @config.details_name.to_sym
+        @details_id = "#{@details_name}_id".to_sym
+        @details_type = "#{@details_name}_type".to_sym
       end
 
       def boolean_eval(t)
@@ -13,16 +16,16 @@ module CanBe
 
       def update_field(t, save = false)
         if save
-          original_details = @model.details
+          original_details = @model.send(@details_name)
           @model.update_attributes(@field_name => t)
-          original_details.destroy unless original_details.class == @model.details.class
+          original_details.destroy unless original_details.class == @model.send(@details_name).class
         else
           self.field_value = t
         end
 
         if block_given?
-          yield(@model.details)
-          @model.details.save if save
+          yield(@model.send(@details_name))
+          @model.send(@details_name).save if save
         end
       end
 
@@ -40,12 +43,12 @@ module CanBe
       end
 
       def initialize_details
-        set_details(field_value.to_sym) if has_details? && !@model.details_id
+        set_details(field_value.to_sym) if has_details? && !@model.send(@details_id)
       end
 
       private
       def has_details?
-        @model.respond_to?(:details) && @model.respond_to?(:details_id) && @model.respond_to?(:details_type)
+        @model.respond_to?(@details_name) && @model.respond_to?(@details_id) && @model.respond_to?(@details_type)
       end
 
       def set_details(t)
@@ -54,10 +57,10 @@ module CanBe
         classname = @config.details[t.to_sym]
 
         if classname
-          @model.details = classname.to_s.camelize.constantize.new
+          @model.send("#{@details_name}=", classname.to_s.camelize.constantize.new)
         else
-          @model.details_id = nil
-          @model.details_type = nil
+          @model.send("#{@details_id}=", nil)
+          @model.send("#{@details_type}=", nil)
         end
       end
     end