+++ spec/support/schema.rb
@@ -17,6 +17,27 @@ ActiveRecord::Schema.define do
     t.timestamps
   end
 
+  create_table :custom_uploads, :force => true do |t|
+    t.string :can_be_type
+    t.integer :custom_details_id
+    t.string :custom_details_type
+    t.timestamps
+  end
+
+  create_table :custom_image_upload_details, :force => true do |t|
+    t.string :format
+    t.timestamps
+  end
+
+  create_table :custom_video_upload_details, :force => true do |t|
+    t.string :encoding
+    t.timestamps
+  end
+
+  create_table :custom_thumbnail_upload_details, :force => true do |t|
+    t.timestamps
+  end
+
   create_table :uploads, :force => true do |t|
     t.string :can_be_type
     t.integer :details_id