+++ lib/can_be/builder/can_be.rb
@@ -100,6 +100,10 @@ module CanBe
           after_initialize do |model|
             model.can_be_processor.initialize_details
           end
+
+          after_save do |model|
+            model.can_be_processor.clean_details
+          end
         end
       end