+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Allowed the options to be specified inside the can_be method call block
 * Added a block processor to each of the `change_to` methods
+* Added the ability to configure the relationship name
 
 ## Version 0.2.1