+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added a block processor to each of the `change_to` methods
 * Added the ability to configure the relationship name
 * Implemented a set of custom RSpec matcher
+* History can be kept to easily get back to data after switching can_be types
 
 ## [Version 0.2.1](https://github.com/mstarkman/can_be/tree/v0.2.1)