+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added the ability to configure the relationship name
 * Implemented a set of custom RSpec matcher
 * History can be kept to easily get back to data after switching can_be types
+* Fixed issue where calling a change_to method didn't destroy the original details record upon save
 
 ## [Version 0.2.1](https://github.com/mstarkman/can_be/tree/v0.2.1)